requirement.txt
absl-py                 0.15.0
astunparse              1.6.3
clang                   5.0
customtkinter           5.2.2
cycler                  0.12.1
darkdetect              0.8.0
flatbuffers             1.12
fonttools               4.54.1
gast                    0.4.0
google-pasta            0.2.0
grpcio                  1.67.0
h5py                    3.1.0
importlib_metadata      8.5.0
joblib                  1.4.2
keras                   2.6.0
keras-facenet           0.3.2
Keras-Preprocessing     1.1.2
kiwisolver              1.4.7
Markdown                3.7
MarkupSafe              3.0.2
matplotlib              3.5.2
mtcnn                   0.1.1
mysql-connector-python  9.1.0
numpy                   1.19.5
opencv-python           4.4.0.46
opt-einsum              3.3.0
packaging               24.1
pillow                  11.0.0
pip                     23.2.1
protobuf                3.20.3
ptl                     0.3.0
pyparsing               3.2.0
python-dateutil         2.9.0.post0
scikit-learn            0.24.1
scipy                   1.6.1
setuptools              68.2.0
six                     1.15.0
tensorboard             2.18.0
tensorboard-data-server 0.7.2
tensorflow              2.6.0
tensorflow-estimator    2.15.0
termcolor               1.1.0
threadpoolctl           3.5.0
typing-extensions       3.7.4.3
Werkzeug                3.0.5
wheel                   0.41.2
wrapt                   1.12.1
zipp                    3.20.2
